Internal: added imageBox and text-path link test
